Rename Server to AIServer.
