+++ src/server/AIServer/Compress.cpp
@@ -1,7 +1,7 @@
 // Compress.cpp : implementation of the CMainFrame class
 //
-#include "stdafx.h"
-//#include "server.h"
+#include "StdAfx.h"
+//#include "AIServer.h"
 #include "implode.h"
 #include "Compress.h"
