+++ src/server/AIServer/MagicTableSet.cpp
@@ -1,8 +1,8 @@
 // MagicTableSet.cpp : implementation file
 //
 
-#include "stdafx.h"
-#include "server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
 #include "MagicTableSet.h"
 
 #ifdef _DEBUG
