+++ src/server/AIServer/MagicType2Set.cpp
@@ -1,8 +1,8 @@
 // MagicType2Set.cpp : implementation file
 //
 
-#include "stdafx.h"
-#include "server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
 #include "MagicType2Set.h"
 
 #ifdef _DEBUG
