+++ src/server/AIServer/MagicType3Set.cpp
@@ -1,8 +1,8 @@
 // MagicType3Set.cpp : implementation file
 //
 
-#include "stdafx.h"
-#include "server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
 #include "MagicType3Set.h"
 
 #ifdef _DEBUG
