+++ src/server/AIServer/MakeItemTableSet.cpp
@@ -1,8 +1,8 @@
 // MakeItemTableSet.cpp : implementation file
 //
 
-#include "stdafx.h"
-#include "server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
 #include "MakeItemTableSet.h"
 
 #ifdef _DEBUG
