+++ src/server/AIServer/MakeWeaponTableSet.cpp
@@ -1,8 +1,8 @@
 // MakeWeaponTableSet.cpp : implementation file
 //
 
-#include "stdafx.h"
-#include "server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
 #include "MakeWeaponTableSet.h"
 
 #ifdef _DEBUG
