+++ src/server/AIServer/Npc.cpp
@@ -1,7 +1,7 @@
-#include "stdafx.h"
+#include "StdAfx.h"
 #include "math.h"
 #include "Npc.h"
-#include "Serverdlg.h"
+#include "AIServerDlg.h"
 #include "Gamesocket.h"
 #include "Region.h"
 #include "Party.h"
