+++ src/server/AIServer/NpcItemSet.cpp
@@ -1,8 +1,8 @@
 // NpcItemSet.cpp : implementation file
 //
 
-#include "stdafx.h"
-#include "server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
 #include "NpcItemSet.h"
 
 #ifdef _DEBUG
