+++ src/server/AIServer/NpcPosSet.cpp
@@ -1,8 +1,8 @@
 // NpcPosSet.cpp : implementation file
 //
 
-#include "stdafx.h"
-#include "server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
 #include "NpcPosSet.h"
 
 #ifdef _DEBUG
