+++ src/server/AIServer/NpcTable.cpp
@@ -2,8 +2,8 @@
 //
 //////////////////////////////////////////////////////////////////////
 
-#include "stdafx.h"
-#include "Server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
 #include "NpcTable.h"
 
 #ifdef _DEBUG
