+++ src/server/AIServer/NpcTableSet.cpp
@@ -1,8 +1,8 @@
 // NpcTableSet.cpp : implementation file
 //
 
-#include "stdafx.h"
-#include "server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
 #include "NpcTableSet.h"
 
 #ifdef _DEBUG
