+++ src/server/AIServer/Party.cpp
@@ -2,9 +2,9 @@
 //
 //////////////////////////////////////////////////////////////////////
 
-#include "stdafx.h"
-#include "server.h"
-#include "ServerDlg.h"
+#include "StdAfx.h"
+#include "AIServer.h"
+#include "AIServerDlg.h"
 #include "Party.h"
 
 #ifdef _DEBUG
