+++ src/server/AIServer/Region.cpp
@@ -2,8 +2,8 @@
 //
 //////////////////////////////////////////////////////////////////////
 
-#include "stdafx.h"
-#include "Server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
 #include "Region.h"
 #include "Npc.h"
 #include "User.h"
