+++ src/server/AIServer/RoomEvent.cpp
@@ -2,10 +2,10 @@
 //
 //////////////////////////////////////////////////////////////////////
 
-#include "stdafx.h"
-#include "server.h"
+#include "StdAfx.h"
+#include "AIServer.h"
+#include "AIServerDlg.h"
 #include "RoomEvent.h"
-#include "ServerDlg.h"
 #include "define.h"
 //#include "Npc.h"
